please add @argoc to nodejs package maint. team
Hello,
I have been working with node on a basic level since 2012, and am getting in to more serious projects with it now. As college faculty and past software developer, I am interested in the administration tasks as well as the fun coding parts. The recent breakdown in the event-stream package management likely spurred your new initiative, and I'd like to be a part of it, in part to keep me involved with node while I continue to develop a depth of knowledge in it.
Thank you for your consideration,
Amelia.