Adding myself to the team
Hey! I saw the call to action on NodeWeekly. 

I've been using and building with Node for years now, and have been contributing to projects on Github. I've been wanting to get involved and help the core community, so this seems like a great place to start!

Thank you for considering me for the team.