Wish to join the team
I agree with the issues raised about old, widely used and unmaintained packages and the need for Node to ensure some security and support. I would like to contribute on issues and security maintenance of these packages and grow with the team and implement the effective solutions the community comes up with.